LayoutTests/platform/wpe/fast/canvas/webgl/shader-with-struct-array-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'gl.createShader')
Tests that program compiling/linking with a struct-array construct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true. Was false.
Some tests failed.

TEST COMPLETE
